fix: restore admin navigation and profile toggle
